Move model creation functions to a TestHelpers subdirectory
This makes it clearer that these functions are for testing only.

Re #26535
