+++ qt/scientific_interfaces/test/ISISReflectometry/Batch/BatchPresenterTest.h
@@ -7,8 +7,8 @@
 #ifndef MANTID_CUSTOMINTERFACES_BATCHPRESENTERTEST_H_
 #define MANTID_CUSTOMINTERFACES_BATCHPRESENTERTEST_H_
 
-#include "../../../ISISReflectometry/Common/ModelCreationHelper.h"
 #include "../../../ISISReflectometry/GUI/Batch/BatchPresenter.h"
+#include "../../../ISISReflectometry/TestHelpers/ModelCreationHelper.h"
 #include "../ReflMockObjects.h"
 #include "MantidAPI/FrameworkManager.h"
 #include "MockBatchView.h"
